Rename handler and processing files in hyp3_process
