+++ modules/hyp3_process/src/hyp3_process/__init__.py
@@ -1,4 +1,4 @@
-from .hyp3_process import (
+from .process import (
     Process,
     HandlerFunction
 )
